添加全局排除标签功能
