+++ src/Pixeval/Controls/Entry/EntryViewModel.cs
@@ -70,7 +70,7 @@ public abstract class EntryViewModel<T>(T entry) : ObservableObject, IDisposable
     public XamlUICommand ShowQrCodeCommand { get; } = EntryItemResources.ShowQRCode.GetCommand(IconGlyph.QRCodeED14);
 
     public XamlUICommand ShowPixEzQrCodeCommand { get; } = EntryItemResources.ShowPixEzQrCode.GetCommand(IconGlyph.Photo2EB9F);
-    
+
     private void GenerateLinkCommandOnExecuteRequested(XamlUICommand sender, ExecuteRequestedEventArgs args)
     {
         UiHelper.ClipboardSetText(AppUri.OriginalString);
