+++ src/Pixeval/Controls/FlyoutContent/BookmarkTagSelector.xaml
@@ -13,7 +13,7 @@
         <StackPanel Spacing="5">
             <CheckBox
                 x:Name="CheckBox"
-                Content="私人"
+                Content="/Misc/BookmarkTagSelectorPrivateCheckBox"
                 IsChecked="{x:Bind IsPrivate, Mode=TwoWay}" />
             <labs:TokenView
                 x:Name="TokenView"
