+++ src/Pixeval/Strings/en-us/SettingsPage.resjson
@@ -85,9 +85,6 @@
     "OverwriteDownloadedFileEntry/Description": "The old file with same name will be overwritten once this setting is turned on",
     "OverwriteDownloadedFileEntry/Header": "Overwrite downloaded files",
     "PerformSignOutButton/Content": "Sign out",
-    "RecommendationItemLimitEntry/Description": "How many works will be fetch in Recommend",
-    "RecommendationItemLimitEntry/Header": "Maximum recommend work",
-    "RecommendationItemLimitNumberBox/PlaceholderText": "[500, 2500]",
     "ReleaseNotesHyperlinkButton/Content": "Update log",
     "ReportBugHyperlinkButton/Content": "File an issue on GitHub",
     "CheckForUpdatesButton/Content": "Check update",
