+++ src/Pixeval/Strings/ru-ru/SettingsPage.resjson
@@ -81,9 +81,6 @@
     "OverwriteDownloadedFileEntry/Description": "Если этот параметр включен, при загрузке файлы будут перезаписываться, если они уже существуют",
     "OverwriteDownloadedFileEntry/Header": "Перезаписывать скачанные файлы",
     "PerformSignOutButton/Content": "Выход из системы",
-    "RecommendationItemLimitEntry/Description": "Максимальное количество иллюстраций, получаемых при использовании функции ежедневных рекомендаций",
-    "RecommendationItemLimitEntry/Header": "Ограничение на количество ежедневных рекомендаций",
-    "RecommendationItemLimitNumberBox/PlaceholderText": "Ограничение на количество рекомендаций [500, 2500]",
     "ReleaseNotesHyperlinkButton/Content": "Примечания к выпуску",
     "ReportBugHyperlinkButton/Content": "Сообщить о проблеме на GitHub",
     "CheckForUpdatesButton/Content": "Проверить обновления",
