+++ src/Pixeval/Strings/zh-cn/SettingsPage.resjson
@@ -73,6 +73,8 @@
     "GitHubRepositoryHyperlinkButton/Content": "GitHub 仓库",
     "ItemsViewLayoutTypeEntry/Description": "选择网格布局则每张图片大小相同；选择流式布局则只有每一行的两端会被对齐",
     "ItemsViewLayoutTypeEntry/Header": "缩略图视图对齐规则",
+    "BlockedTagsEntry/Description": "包含屏蔽标签的作品在应用内不会显示",
+    "BlockedTagsEntry/Header": "屏蔽标签",
     "ImageMirrorServerEntry/Description": "下载图片时所使用的的镜像站的主机名，如果不明白是什么意思请留空",
     "ImageMirrorServerEntry/Header": "图片镜像站的主机名",
     "ImageMirrorServerTextBox/PlaceholderText": "主机名",
@@ -100,9 +102,6 @@
     "PerformSignOutButton/Content": "执行登出",
     "DownloadWhenBookmarkedEntry/Description": "在收藏作品后，会自动下载该作品",
     "DownloadWhenBookmarkedEntry/Header": "收藏作品时下载",
-    "RecommendationItemLimitEntry/Description": "使用每日推荐功能时最多会获取多少条插画",
-    "RecommendationItemLimitEntry/Header": "每日推荐结果数量限制",
-    "RecommendationItemLimitNumberBox/PlaceholderText": "每日推荐数量限制[500, 2500]",
     "ReleaseNotesHyperlinkButton/Content": "更新日志",
     "ReportBugHyperlinkButton/Content": "在 GitHub 上汇报问题",
     "CheckForUpdatesButton/Content": "检查更新",
